requirements.txt
# Core dependencies
boto3>=1.34.0
pyyaml>=6.0
rich>=13.7.0

# AgentCore SDK
bedrock-agentcore-sdk-python>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.0

# Strands Agents
strands-agents>=0.1.0
strands-tools>=0.1.0

# Additional utilities
requests>=2.31.0
python-dotenv>=1.0.0
